+++ pages/_app.js
@@ -1,7 +1,4 @@
 import '../styles/Shark.css'
-import '../styles/bootstrap/bootstrap-grid.css'
-import '../styles/bootstrap/bootstrap-reboot.css'
-import '../styles/bootstrap/bootstrap-utilities.css'
 
 function MyApp({ Component, pageProps }) {
   return <Component {...pageProps} />