+++ pages/post/[name].jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import Layout from '../../utils/Modules/Layout';
 import style from '../../styles/PostIndex.module.css';
 import Router from 'next/router';
+import { Container } from '@material-ui/core';
 
 const index = () => {
 
@@ -10,7 +11,7 @@ const index = () => {
     }
 
     return (
-        <Layout>
+        <Layout maxWidth="lg">
             <div className={style.title}>
                 <img className={style.back} onClick={back} src="/icons/arrow-left-solid.svg" height="28px" width="28px" />
                 <div className={style.vhr} />
@@ -24,7 +25,6 @@ const index = () => {
 
                 <div className={style.content}>xdd</div>
             </div>
-
         </Layout>
     );
 };